Modify each function that accepts Params to call to_np on the params.
Params that have been saved to json then loaded typically come back as a nested dict of lists, when what we really want is a nested dict of numpy arrays. It's also possible to end up with a nested dict of jax-numpy arrays which we also do not want because they are slow to work with. Instead of checking and throwing an error it's easier to just convert these types to the type that we want.

PiperOrigin-RevId: 839439298