Fix issue with jitting `step_network`
The `to_np` call recently added to `rnn_utils.step_network` is convenient when the RNN params are, e.g., saved in JSON, and need to be converted to numpy dicts before use. Unfortunately, it also prevents `step_network` from being jittable. In this CL, we make this call optional (but default to True) so that it can be disabled. Anyone who needs to jit `step_network` will have to call `to_np` themselves on the params before passing to `step_network`.

PiperOrigin-RevId: 843677095